Debug stuff. Axis indicator working.

+++ Source/Source/pick.cpp
@@ -70,18 +70,18 @@ void raycast(glm::vec3 origin, glm::vec3 direction, float radius, std::function<
 	float dy = direction[1];
 	float dz = direction[2];
 	// Direction to increment x,y,z when stepping.
-	int stepX = signum(dx);
-	int stepY = signum(dy);
-	int stepZ = signum(dz);
+	float stepX = signum(dx);
+	float stepY = signum(dy);
+	float stepZ = signum(dz);
 	// See description above. The initial values depend on the fractional
 	// part of the origin.
 	float tMaxX = intbound(origin[0], dx);
 	float tMaxY = intbound(origin[1], dy);
 	float tMaxZ = intbound(origin[2], dz);
 	// The change in t when taking a step (always positive).
-	int tDeltaX = stepX / dx;
-	int tDeltaY = stepY / dy;
-	int tDeltaZ = stepZ / dz;
+	float tDeltaX = stepX / dx;
+	float tDeltaY = stepY / dy;
+	float tDeltaZ = stepZ / dz;
 	// Buffer for reporting faces to the callback.
 	glm::vec3 face(0); // probably needs to point in the direction it faces
 
@@ -137,7 +137,8 @@ void raycast(glm::vec3 origin, glm::vec3 direction, float radius, std::function<
 		}
 		else
 		{
-			if (tMaxY < tMaxZ) {
+			if (tMaxY < tMaxZ)
+			{
 				if (tMaxY > radius) break;
 				y += stepY;
 				tMaxY += tDeltaY;
@@ -159,3 +160,114 @@ void raycast(glm::vec3 origin, glm::vec3 direction, float radius, std::function<
 		}
 	}
 }
+
+// Bresenham3D
+//
+// A slightly modified version of the source found at
+// http://www.ict.griffith.edu.au/anthony/info/graphics/bresenham.procs
+// Provided by Anthony Thyssen, though he does not take credit for the original implementation
+//
+// It is highly likely that the original Author was Bob Pendelton, as referenced here
+//
+// ftp://ftp.isc.org/pub/usenet/comp.sources.unix/volume26/line3d
+//
+// line3d was dervied from DigitalLine.c published as "Digital Line Drawing"
+// by Paul Heckbert from "Graphics Gems", Academic Press, 1990
+//
+// 3D modifications by Bob Pendleton. The original source code was in the public
+// domain, the author of the 3D version places his modifications in the
+// public domain as well.
+//
+// line3d uses Bresenham's algorithm to generate the 3 dimensional points on a
+// line from (x1, y1, z1) to (x2, y2, z2)
+
+
+//void Bresenham3D(int x1, int y1, int z1, const int x2, const int y2, const int z2, 
+//	std::function<bool(float, float, float, BlockPtr, glm::vec3)> callback)
+//{
+//	int i, dx, dy, dz, l, m, n, x_inc, y_inc, z_inc, err_1, err_2, dx2, dy2, dz2;
+//	int point[3];
+//
+//	point[0] = x1;
+//	point[1] = y1;
+//	point[2] = z1;
+//	dx = x2 - x1;
+//	dy = y2 - y1;
+//	dz = z2 - z1;
+//	x_inc = (dx < 0) ? -1 : 1;
+//	l = abs(dx);
+//	y_inc = (dy < 0) ? -1 : 1;
+//	m = abs(dy);
+//	z_inc = (dz < 0) ? -1 : 1;
+//	n = abs(dz);
+//	dx2 = l << 1;
+//	dy2 = m << 1;
+//	dz2 = n << 1;
+//
+//	if ((l >= m) && (l >= n))
+//	{
+//		err_1 = dy2 - l;
+//		err_2 = dz2 - l;
+//		for (i = 0; i < l; i++)
+//		{
+//			//output->getTileAt(point[0], point[1], point[2])->setSymbol(symbol);
+//			callback(x, y, z, Chunk::AtWorld(glm::ivec3(x, y, z)), face);
+//			if (err_1 > 0)
+//			{
+//				point[1] += y_inc;
+//				err_1 -= dx2;
+//			}
+//			if (err_2 > 0)
+//			{
+//				point[2] += z_inc;
+//				err_2 -= dx2;
+//			}
+//			err_1 += dy2;
+//			err_2 += dz2;
+//			point[0] += x_inc;
+//		}
+//	}
+//	else if ((m >= l) && (m >= n))
+//	{
+//		err_1 = dx2 - m;
+//		err_2 = dz2 - m;
+//		for (i = 0; i < m; i++)
+//		{
+//			output->getTileAt(point[0], point[1], point[2])->setSymbol(symbol);
+//			if (err_1 > 0) {
+//				point[0] += x_inc;
+//				err_1 -= dy2;
+//			}
+//			if (err_2 > 0)
+//			{
+//				point[2] += z_inc;
+//				err_2 -= dy2;
+//			}
+//			err_1 += dx2;
+//			err_2 += dz2;
+//			point[1] += y_inc;
+//		}
+//	}
+//	else
+//	{
+//		err_1 = dy2 - n;
+//		err_2 = dx2 - n;
+//		for (i = 0; i < n; i++)
+//		{
+//			output->getTileAt(point[0], point[1], point[2])->setSymbol(symbol);
+//			if (err_1 > 0) {
+//				point[1] += y_inc;
+//				err_1 -= dz2;
+//			}
+//			if (err_2 > 0)
+//			{
+//				point[0] += x_inc;
+//				err_2 -= dz2;
+//			}
+//			err_1 += dy2;
+//			err_2 += dx2;
+//			point[2] += z_inc;
+//		}
+//	}
+//	output->getTileAt(point[0], point[1], point[2])->setSymbol(symbol);
+//}
